requirements.txt
Django==1.7.4
django-appconf==0.6
django-compressor==1.4
django-crispy-forms==1.4.0
django-haystack==2.3.1
django-pagedown==0.1.0
django-watson==1.1.5
Markdown==2.5.2
PyMySQL==0.6.3
six==1.9.0
Whoosh==2.6.0
